Persistent, size-limited cache for bdd_apply operations.
This new cache is the sort used in other BDD packages, such as BuDDy
and CUDD.  Rather than cache EVERY computed value, which requires
either an enormous array or hash table with buckets, a fixed-size,
hashed array is used instead:  looking up the entry for a key never
returns a null pointer, but may return an invalid cache entry, which
needs to be rewritten.

This caching scheme has a much lower constant factor than the regular
hash table-based cache, and with a reasonable hashing function,
recomputation of previously invalidated cache entries seems to not be
a big problem.