Big changes to support garbage collection; big refactoring.
This should have been done in smaller changesets.  Oh well.  An
overview of the changes:

    * 'bool' now renamed 'boolean' and moved to its own file.

    * Now there are two types of BDDs: user-level BDDs and raw BDDs,
      with a one-to-one mapping between the user-level BDDs and a
      subset of the raw BDDs.  This extra level of indirection will
      allow garbage collection with compaction of the BDD graph, which
      requires that raw BDDs be shuffled around.

    * Two new instances of the generic hash table have been added,
      both to support the user-level/raw BDD mapping.  This hash table
      implementation could be better, e.g., by using memory pools and
      by exposing the implementation to allow inlining.

    * bdd.h renamed to bddlib.h.  I need a good name for this library.

    * bdd_true and bdd_false are now functions that take a bdd_mgr_t
      pointer, rather than global constant values.

    * The Haskell bindings are broken.

I think this extra level of indirection is somewhat novel, although
I'm not certain.  BuDDY does not use this sort of indirection, and I
don't believe CUDD does either.  The number of user-level BDDs will
usually be much less than the number of raw BDDs, but only user-level
BDDs need to be reference counted, and so the space savings can be
large.  E.g., for the 12-queens puzzle, some 25M raw BDDs are created,
but less than 7000 user-level BDDs are; In BuDDy, each node requires 1
byte for reference counting, and so 25 MB would be needed, whereas
with the user/raw BDD distinction, even at an entire int per BDD, less
than 60 KB would be needed.