Eliminated auto-gc-on-deref; added automatic GC when out of nodes.
I was avoiding this before, since interrupting a BDD operation to do
GC could result in an endless loop if the operation requires more
nodes than the heap capacity.  To fix this shortcoming, at each
out-of-nodes handler in the code, we keep track of whether the
operation has already been interrupted to do GC since last resizing,
and if so, we don't attempt GC again.

This exception handling code is low-level, error-prone, and ugly.
Fortunately it is only used in three places, around where
_bdd_make_node is called.