Checkpointing buggy variable swapping code.
I think the flaw with this code is that it tries to do in-place node
updates, which are impossible with the hashing scheme I am using
(which is linear probing, using the node vector also as the node hash
table).  Changing the fields of a node changes its hash code, and if
the node is left where it is, everything breaks.  Instead of doing
in-place updates, I would need to create new nodes, which would
potentially require copying the entire BDD graph to have all nodes
point into the new locations.  Uggg.