New hash table.
This hash table is kept separate from the nodes array.  Separate
chaining is used, with a max load of 2.  With this implementation,
each node requires 10 bytes (with sizeof(unsigned) == 4) for the hash
table, as opposed to 4 bytes with the previous, non-working scheme.

This implementation also includes support for doubling the capacity of
the manager.  It should furthermore be possible to simply implement the
required hash table updates required by in-place node array updates
done during variable swapping.