Reference counting re-implemented.
God, this low-level coding is brutal, and I don't trust my code.

All the existing practical BDD packages that I know of use a
combination of manual reference counting & garbage collection to allow
for reuse of parts of the BDD data structures; the user is responsible
for incrementing and decrementing reference counts appropriately, and
when the manager is out of free nodes, a garbage collection frees
nodes with reference counts of zero.

This hybrid reference counting + GC strategy is a bit hokey, as it
prevents deterministic deallocation that is allowed by pure reference
counting, and it requires additional memory per object needed to
implement reference counting which is not needed in pure GC settings.
In a way, it is the worst of both worlds.

Pure GC---i.e., reference counting only for user formulae---seems like
it won't allow for efficient variable reordering, which needs to
perform incremental garbage collection as it goes along.  (However,
the *algorithmic* complexity of doing non-incremental GC when swapping
may be the same; I have to think more about it.)

What has stopped others from using pure reference counting seems to be
the computed result cache (i.e., ite_cache)---if a node is referenced
by no other nodes and by no user formulae, it may still be contained
in the computed result cache, and it is O(size of cache) to make sure
this isn't the case, and so the check is postponed until the GC phase.

The code in this revision implements pure reference counting, with no
postponed GC phase.  I use an unsigned for the reference count for
each node; this reference count is the sum of references from user
formulae, other nodes, and entries in the computed result cache.  A
reference count of zero means that node is available for use.

Pros of this implementation:
    * Deallocation is deterministic and not postponed.
    * No need for longjmp: when out-of-nodes, the manager can grow
      mid-operation without anything complicated.
    * Growing the manager is simpler:  can use realloc; no need to
      clear the computed result cache.
    * Implementation seems simpler than other approaches, as there is
      no need to do node reclamations or worry about non-local jumps.
    * Incremental collection should be easy during reordering.

Cons of this implementation:
    * A node may live longer than it ought to if the only references
      to it are in the computed result cache.  (However, with the
      cache replacement rates I've observed, it looks like this won't
      be a significant problem.)

What a commit message!