Fixed a reference counting bug.
Note, the low & high arguments to _bdd_make_node must correspond to
nodes with non-zero reference counts.  The old code didn't always do
this, and in certain cases the node at low or high would be rewritten
with a different node.  In the case of the queens problem with n=8,
this would cause a cycle in the BDD graph, resulting in a stack
overflow...

Perhaps all the BDD library functions that return bdd_t *should*
implicitly increment the reference count.  Although I don't know that
this would make the code easier for me to write correctly.