Eliminate the _bdd_mgr_check_invariants call in _bdd_mgr_double_capacity.
The invariants do not hold when a doubling occurs while swapping
variables around.  In particular, if a swapping is midway through and
a doubling occurs, it is possible for mgr->nodes[i] to equal
mgr->nodes[j] with i < j, if the doubling occurs before mgr->nodes[j]
has been swapped.