Updated the Haskell bindings.
These bindings are incomplete.  Some thought it required to come up
with useful, usable bindings.  The bindings must allow for timely
dereferencing of unused BDDs for reasonable memory performance.  At
the same time the interface should be as simple as possible and
statically prevent errors as much as is feasible.

I am currently using a phantom type parameter in the Mgr and Bdd
type constructors, and have given the function to create a new manager
an existential type

    newMgrWithHint :: Int -> Int -> (forall s. Mgr s -> IO a) -> IO a,

which statically ensures that only BDDs from a given manager are used
with that manager, and that no BDDs once the manager is destroyed.

Unfortunately, my first attempt at timely BDD dereferencing using FFI
finalizers proved incorrect: sometimes the finalizer on a BDD (which
decrements the reference count of the BDD for its manager) would be
run after the manager's finalizer (which frees the manager) was run,
resulting in a segfault.  Worse, there is no good way to ensure that
the finalizer for one object is called after the finalizer for
another.

My second attempt using 'Control.Exception.finally' is equally
unsatisfying: it requires that programs using the library be twisted
into semi-CPS style, and BDDs can be leaked outside of the region in
which they live.

One possible solution is to use reference counting on the BDD manager
type in Haskell land.  Instead of having 'Raw.BddMgr' be a pointer to
a 'bdd_mgr_t', we'd make it point to a counted reference such as

    struct mgr_ref_t {
        bdd_mgr_t *mgr;
        unsigned ref_cnt;
    },

which would have its ref_cnt started at 1 and incremented each time
'bdd_inc_ref' is called with the manager, and would have its ref_cnt
decremented whenever a finalizer is called on the manager or one of
its BDDs.  The manager's finalizer would still free the 'mgr' field.
The finalizer that drops 'ref_cnt' to 0 frees the structure.  (It's a
simple C version of a shared pointer.)  This would avoid the problem
of requiring an ordering among finalizers.